+++ src/components/Catalog/CatalogItem.scss
