+++ src/components/layout/layout.tsx
@@ -1,4 +1,4 @@
-import './layout.scss';
+import './Layout.scss';
 import { Outlet } from 'react-router-dom';
 import { Header } from '../Header/Header';
 import { Aside } from '../Aside/Aside';
